vertex_ai_rag/pyproject.toml
[project]
name = "vertex-ai-rag"
version = "0.1.0"
description = "Vertex AI RAG implementation for client knowledge management"
authors = [
    {name = "Vectara Team", email = "team@vectara.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]==1.108.0",
    "google-adk==1.10.0",
    "google-auth>=2.35.0",
    "google-api-python-client>=2.154.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "openai>=1.99.3",
    "pydantic>=2.11.5",
    "beautifulsoup4>=4.12.3"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "black>=24.10.0",
    "isort>=5.13.2",
    "mypy>=1.13.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

